extended model compatibility by changing mechanic for model bean instatiation
