+++ src/main/java/es/thalesalv/gptbot/application/service/usecases/ChatbotUseCase.java
@@ -13,7 +13,7 @@ import org.springframework.stereotype.Component;
 import es.thalesalv.gptbot.adapters.data.ContextDatastore;
 import es.thalesalv.gptbot.application.config.Persona;
 import es.thalesalv.gptbot.application.service.ModerationService;
-import es.thalesalv.gptbot.application.service.models.gpt.GptModel;
+import es.thalesalv.gptbot.application.service.interfaces.GptModel;
 import es.thalesalv.gptbot.application.util.MessageUtils;
 import lombok.RequiredArgsConstructor;
 import net.dv8tion.jda.api.entities.Mentions;
